test(meshtastic): redistribute edge case tests into domain-specific suites
Move all test cases from `test_meshtastic_utils_edge_cases.py` into their respective functional test files to improve organization and maintainability.

- BLE tests moved to `test_meshtastic_utils_ble.py`
- Connection and retry tests moved to `test_meshtastic_utils_connect.py`
- Serial port validation tests moved to `test_meshtastic_utils_connect_paths.py`
- Disconnection and reconnection tests moved to `test_meshtastic_utils_disconnect.py`
- Message processing and malformed packet tests moved to `test_meshtastic_utils_message_edge.py`
- Reply and database error tests moved to `test_meshtastic_utils_messages.py`
- Routing and plugin timeout tests moved to `test_meshtastic_utils_messages_routing.py`
- Service detection tests moved to `test_meshtastic_utils_service.py`

`test_meshtastic_utils_edge_cases.py` is retained as a documentation placeholder.
